requirements.txt
fastapi>=0.135.1
uvicorn[standard]>=0.42.0
python-multipart>=0.0.22
zxing-cpp>=3.0.0
Pillow>=12.1.1
pdf2image>=1.17.0
google-auth>=2.29.0
pydantic-settings>=2.0.0
